requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.8.2
pydantic-settings==2.3.4
SQLAlchemy==2.0.34
python-multipart==0.0.9
pillow==10.4.0
openai==1.40.0
httpx==0.27.0
pillow-heif==0.16.0; python_version < "3.13"
transformers==4.46.1
accelerate==0.33.0
torch==2.7.1
tokenizers==0.20.3
easyocr==1.7.1
opencv-python-headless==4.10.0.84

